Add files via upload
In this project, we aim to harness the power of Natural Language Processing (NLP) techniques to analyze and understand the sentiments expressed in Amazon product reviews. Leveraging advanced machine learning algorithms and libraries such as TensorFlow and scikit-learn, we delve into a dataset consisting of thousands of Amazon product reviews spanning various categories. Our goal is to develop a robust sentiment analysis model capable of accurately classifying these reviews into positive, negative, or neutral sentiments. Throughout the project, we explore different methodologies for data preprocessing, feature extraction, and model training, striving to achieve optimal performance and interpretability. Additionally, we employ visualization techniques to gain insights into the distribution of sentiments across different products and categories. By the end of this endeavor, we aim to create a comprehensive sentiment analysis framework that not only provides valuable insights for businesses but also enhances our understanding of customer sentiments in the digital age.